+++ PackageEncapsulation.cpp
@@ -14,7 +14,7 @@ std::vector<char> PackageEncapsulation::toBytes(std::unique_ptr<Message> msg)
     std::vector<char> byteVec;
     byteVec = toBytes<uint8_t>(static_cast<uint8_t>(msg->payload().size()));
     res.insert(res.end(), byteVec.begin(), byteVec.end());
-    byteVec = toBytes<uint16_t>(msg->seqNumber());
+    byteVec = toBytes<uint32_t>(msg->seqNumber());
     res.insert(res.end(), byteVec.begin(), byteVec.end());
     byteVec = toBytes<uint16_t>(msg->ackNumber());
     res.insert(res.end(), byteVec.begin(), byteVec.end());
@@ -52,7 +52,7 @@ std::unique_ptr<Message> PackageEncapsulation::toMessage(const std::vector<char>
                 std::vector<char>(mAccumulatingVector.begin()+HEADER_LEN+LENGTH_LEN+SEQ_NUM_LEN+ACK_NUM_LEN,
                                   mAccumulatingVector.begin()+HEADER_LEN+LENGTH_LEN+SEQ_NUM_LEN+ACK_NUM_LEN+SENSORTYPE_LEN)
                 );
-    uint16_t seqNumber = toValue<uint16_t>(
+    uint16_t seqNumber = toValue<uint32_t>(
                 std::vector<char>(mAccumulatingVector.begin()+HEADER_LEN+LENGTH_LEN,
                                   mAccumulatingVector.begin()+HEADER_LEN+LENGTH_LEN+SEQ_NUM_LEN)
                 );
